Develop

+++ src/main/java/fr/tobby/tripnjoyback/service/ProfileService.java
@@ -39,12 +39,19 @@ public class ProfileService {
     public ProfileModel createUserProfile(long userId, ProfileCreationRequest profileCreationRequest)
     {
         setProfileInactive(userId);
+        ProfileModel createdProfileModel = createProfile(profileCreationRequest);
+        userRepository.findById(userId).orElseThrow(UserNotConfirmedException::new)
+                      .getProfiles().add(profileRepository.getById(createdProfileModel.getId()));
+        return createdProfileModel;
+    }
+
+    @Transactional
+    public ProfileModel createProfile(ProfileCreationRequest profileCreationRequest)
+    {
         ProfileEntity profileEntity = ProfileEntity.builder()
                                                    .name(profileCreationRequest.getName())
                                                    .active(true).build();
-        userRepository.findById(userId).orElseThrow(UserNotConfirmedException::new)
-                      .getProfiles().add(profileEntity);
-        profileRepository.save(profileEntity);
+        profileEntity = profileRepository.save(profileEntity);
         AnswersEntity answersEntity = createAnswersEntity(profileCreationRequest, profileEntity.getId());
         return ProfileModel.of(profileEntity, answersEntity);
     }
@@ -144,6 +151,8 @@ public class ProfileService {
     public void deleteProfile(long profileId)
     {
         ProfileEntity profileEntity = profileRepository.findById(profileId).orElseThrow(() -> new ProfileNotFoundException("No profile with this id"));
+        if (profileEntity.isActive())
+            throw new IllegalArgumentException("Cannot delete an active profile");
         profileRepository.delete(profileEntity);
         AnswersEntity answersEntity = answersRepository.findByProfileId(profileId);
         answersRepository.deleteByProfileId(answersEntity.getProfileId());
